--- a/pyiron_atomistics/lammps/units.py
+++ b/pyiron_atomistics/lammps/units.py
@@ -60,6 +60,7 @@ LAMMPS_UNIT_CONVERSIONS = {
         "temperature": 1.0,
         "pressure": GPA_TO_BAR,
         "charge": 1.0,
+        "natoms": 1,
     },
     "si": {
         "mass": AMU_TO_KG,
@@ -71,6 +72,7 @@ LAMMPS_UNIT_CONVERSIONS = {
         "temperature": 1.0,
         "pressure": GPA_TO_PA,
         "charge": EL_TO_COUL,
+        "natoms": 1,
     },
     "cgs": {
         "mass": AMU_TO_G,
@@ -82,6 +84,7 @@ LAMMPS_UNIT_CONVERSIONS = {
         "temperature": 1.0,
         "pressure": GPA_TO_BARYE,
         "charge": 4.8032044e-10,  # In statCoulombs, but these are deprecated and thus not in scipt.constants
+        "natoms": 1,
     },
     "real": {
         "mass": 1.0,
@@ -93,6 +96,7 @@ LAMMPS_UNIT_CONVERSIONS = {
         "temperature": 1.0,
         "pressure": GPA_TO_ATM,
         "charge": 1.0,
+        "natoms": 1,
     },
     "electron": {
         "mass": 1.0,
@@ -104,6 +108,7 @@ LAMMPS_UNIT_CONVERSIONS = {
         "temperature": 1.0,
         "pressure": GPA_TO_PA,
         "charge": 1.0,
+        "natoms": 1,
     },
 }
 
@@ -136,6 +141,7 @@ _conversion_dict["mass"] = ["mass"]
 _conversion_dict["charge"] = ["charges", "charge"]
 _conversion_dict["force"] = ["forces", "force", "mean_forces"]
 _conversion_dict["dimensionless_integer_quantity"] = ["steps", "indices"]
+_conversion_dict["natoms"] = ["natoms"]
 
 # Reverse _conversion_dict
 quantity_dict = dict()