Fixes #671 for my calculations at least. Temporarily I fixed it by allocating 26GB of RAM for my notebooks, but even that was not sufficient for some of my recent calculations.
With this one of the simulations that kept killing the kernel during parsing never allocated more than 3-4 GB of RAM (according to top, didnt check in detail)

Did not check whether this gives the same numbers as the old parser yet. Is correct parsing part of unit tests?