@pmrv 2022-08-09
For testing there's some water trajectories that are checked in a round-about way, but I would have to see if it applies here.  I would just move the old method in the unit tests and include a sample dump in the test resources, then check if new and old give the same results.