@Leimeroth 2022-08-09
> For testing there's some water trajectories that are checked in a round-about way, but I would have to see if it applies here. I would just move the old method in the unit tests and include a sample dump in the test resources, then check if new and old give the same results.

The water tests failed and seems that everything is covered, but locally the integration tests ran now, so I assume I fixed the last problems. Adding the old method to the unit tests would require to permanently keep it in. I wouldn't mind a little more code, but maybe it would be confusing or is not wanted to not have too much old struff in there.