@Leimeroth 2022-08-25
I did some benchmarks regarding timing and memory usage (using ipythons %%timeit and %%memit requiring the memory_profiler package) on job.collect_dump_file() and job.collect_dump_file_old().
The new version is about 1.2 times faster and more important, it requires 1/3 to 1/2 of the memory. While %%timeit only has a very low error the memory usage changes a lot more from test to test. f.e.
new:
peak memory: 673.71 MiB, increment: 340.40 MiB
peak memory: 647.71 MiB, increment: 314.33 MiB
peak memory: 616.68 MiB, increment: 283.29 MiB
peak memory: 580.42 MiB, increment: 247.03 MiB
peak memory: 598.57 MiB, increment: 265.18 MiB
peak memory: 611.34 MiB, increment: 277.95 MiB
peak memory: 673.91 MiB, increment: 340.53 MiB
peak memory: 598.67 MiB, increment: 265.28 MiB
2.96 s ± 14 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

Old:
peak memory: 1104.89 MiB, increment: 771.50 MiB
peak memory: 1097.41 MiB, increment: 673.98 MiB
peak memory: 1179.28 MiB, increment: 749.60 MiB
peak memory: 1200.61 MiB, increment: 773.00 MiB
peak memory: 1125.32 MiB, increment: 697.46 MiB
peak memory: 1201.58 MiB, increment: 773.46 MiB
peak memory: 1201.09 MiB, increment: 762.42 MiB
peak memory: 1183.31 MiB, increment: 744.39 MiB
3.57 s ± 36.8 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)